GraphMakie

Documentation for GraphMakie.

GraphMakie.closest_point

GraphMakie.graphplot

GraphMakie.graphplot!

GraphMakie.closest_point — Method

closest_point(p, positions)

Returns the index of the point in positions closest to p

GraphMakie.graphplot! — Method

graphplot(graph::AbstractGraph)

Creates a plot of the network graph. Consists of multiple steps:

Layout the nodes: the layout attribute is has to be a function f(adj_matrix)::pos where pos is either an array of Point2f0 or (x, y) tuples

plot edges as linesegments-plot

plot nodes as scatter-plot

The main attributes for the subplots are exposed as attributes for graphplot. Additional attributes for the scatter or linesegments plots can be provided as a named tuples to node_attr and edge_attr.

Attributes

Available attributes and their defaults for Combined{GraphMakie.graphplot!, T} where T are:

GraphMakie.graphplot — Method

graphplot(graph::AbstractGraph)

Creates a plot of the network graph. Consists of multiple steps:

Layout the nodes: the layout attribute is has to be a function f(adj_matrix)::pos where pos is either an array of Point2f0 or (x, y) tuples

plot edges as linesegments-plot

plot nodes as scatter-plot

The main attributes for the subplots are exposed as attributes for graphplot. Additional attributes for the scatter or linesegments plots can be provided as a named tuples to node_attr and edge_attr.

Attributes

Available attributes and their defaults for Combined{GraphMakie.graphplot, T} where T are:

  edge_attr    Attributes with 0 entries
  edge_color   :black
  edge_width   1.0
  layout       NetworkLayout.Spring.layout
  node_attr    Attributes with 0 entries
  node_color   :gray65
  node_marker  Circle{T} where T
  node_size    10